+++ packages/core/src/components/ozon-content/ozon-content.tsx
@@ -44,12 +44,6 @@ export class OzonContent implements ComponentInterface {
   inline = false;
 
   /**
-   * Adds a non breaking space to the node element.
-   */
-  @Prop({ reflect: true })
-  addSpaceBeforeNode = false;
-
-  /**
    * To mark text.
    */
   @Prop()
@@ -88,7 +82,6 @@ export class OzonContent implements ComponentInterface {
   render(): JSX.Element {
     const context: OzonContentContext = {
       state: this.state,
-      addSpaceBeforeNode: this.addSpaceBeforeNode,
       inline: this.inline,
       mark: this.mark,
       setState: (state) => (this.state = state),